refactor feedback 적용
